# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Auction Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: getTeams
      tags: [teams]
      summary: List all teams
      parameters:
        - name: conference
          in: query
          schema:
            type: string
            enum: [AFC, NFC]
        - name: division
          in: query
          schema:
            type: string
            enum: [East, North, South, West]
        - name: search
          in: query
          schema:
            type: string
        - name: bidderId
          in: query
          schema:
            type: ["number", "null"]
        - name: season
          in: query
          schema:
            type: ["number", "null"]
        - name: calcuttaId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a single team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Team detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      description: Administrative bearer token supplied through ADMIN_API_KEY.
    JobRunnerBearerAuth:
      type: http
      scheme: bearer
      description: Restricted job trigger token supplied through JOB_RUNNER_SECRET.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Team:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        conference:
          type: string
          enum: [AFC, NFC]
        division:
          type: string
          enum: [East, North, South, West]
        bidAmount:
          type: number
        owners:
          type: array
          items:
            $ref: "#/components/schemas/TeamOwner"
      required: [id, name, conference, division, bidAmount, owners]
    TeamOwner:
      type: object
      properties:
        bidderId:
          type: number
        bidderName:
          type: string
        ownershipShare:
          type: number
      required: [bidderId, bidderName, ownershipShare]
...
    TeamInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        conference:
          type: string
          enum: [AFC, NFC]
        division:
          type: string
          enum: [East, North, South, West]
        bidAmount:
          type: number
          minimum: 0
        season:
          type: number
        calcuttaId:
          type: number
        owners:
          type: array
          items:
            $ref: "#/components/schemas/OwnerInput"
          minItems: 1
      required: [name, conference, division, bidAmount, owners]
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        conference:
          type: string
          enum: [AFC, NFC]
        division:
          type: string
          enum: [East, North, South, West]
        bidAmount:
          type: number
          minimum: 0
        season:
          type: number
        calcuttaId:
          type: number
        owners:
          type: array
          items:
            $ref: "#/components/schemas/OwnerInput"
          minItems: 1
    OwnerInput:
      type: object
      properties:
        bidderId:
          type: number
        ownershipShare:
          type: number
      required: [bidderId, ownershipShare]

    # Results (Calcutta Returns format)
...