# lib/api-zod/src/generated/types/teamInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NFL Auction Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { OwnerInput } from './ownerInput';
import type { TeamInputConference } from './teamInputConference';
import type { TeamInputDivision } from './teamInputDivision';

export interface TeamInput {
  /** @minLength 1 */
  name: string;
  conference: TeamInputConference;
  division: TeamInputDivision;
  /** @minimum 0 */
  bidAmount: number;
  season?: number;
  calcuttaId?: number;
  /** @minItems 1 */
  owners: OwnerInput[];
}