# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NFL Auction Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { OwnerInput } from './ownerInput';
import type { TeamUpdateConference } from './teamUpdateConference';
import type { TeamUpdateDivision } from './teamUpdateDivision';

export interface TeamUpdate {
  /** @minLength 1 */
  name?: string;
  conference?: TeamUpdateConference;
  division?: TeamUpdateDivision;
  /** @minimum 0 */
  bidAmount?: number;
  season?: number;
  calcuttaId?: number;
  /** @minItems 1 */
  owners?: OwnerInput[];
}